Module osauth::sync

Synchronous wrapper for a session.

This module is only available when the sync feature is enabled.

Structs

SyncBody
A synchronous body that can be used with asynchronous code.

SyncSession
A synchronous wrapper for an asynchronous session.

SyncStream
A reader into an asynchronous stream.

Type Definitions

Result
A result of an OpenStack operation.

SyncStreamItem
An item in a SyncStream.